+++ resources/lang/sk/harvest.php
@@ -6,12 +6,14 @@ return [
         'started' => 'Začaté',
         'progress' => 'Spracováva sa (:current/:total)',
         'error' => ':error',
+        'finished' => 'Spracovaných záznamov: :processed',
         'completed' =>
-            'Spracovaných bolo :processed záznamov.' . PHP_EOL .
+            'Spracovaných záznamov: :processed' . PHP_EOL .
             ':created nových záznamov' . PHP_EOL .
             ':updated bolo upravených' . PHP_EOL .
             ':deleted bolo zmazaných' . PHP_EOL .
-            ':skipped bolo preskočených.' . PHP_EOL .
+            ':skipped bolo preskočených' . PHP_EOL .
+            ':failed zlyhalo.' . PHP_EOL .
             'Trvalo to :time s',
     ],
 ];